Move philips_js coordinator to separate module
