+++ .coveragerc
@@ -1028,6 +1028,7 @@ omit =
     homeassistant/components/permobil/entity.py
     homeassistant/components/permobil/sensor.py
     homeassistant/components/philips_js/__init__.py
+    homeassistant/components/philips_js/coordinator.py
     homeassistant/components/philips_js/light.py
     homeassistant/components/philips_js/media_player.py
     homeassistant/components/philips_js/remote.py
