+++ homeassistant/components/philips_js/binary_sensor.py
@@ -13,7 +13,8 @@ from homeassistant.components.binary_sensor import (
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import PhilipsTVConfigEntry, PhilipsTVDataUpdateCoordinator
+from . import PhilipsTVConfigEntry
+from .coordinator import PhilipsTVDataUpdateCoordinator
 from .entity import PhilipsJsEntity
