+++ homeassistant/components/philips_js/entity.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import PhilipsTVDataUpdateCoordinator
+from .coordinator import PhilipsTVDataUpdateCoordinator
 
 
 class PhilipsJsEntity(CoordinatorEntity[PhilipsTVDataUpdateCoordinator]):
